Fix 32-bit windows test failure: 2^32 overflows

+++ test/fldmod_by_const.jl
@@ -2,7 +2,7 @@ using Compat.Test
 using FixedPointDecimals
 
 @testset "fldmod_by_const" begin
-    vals = [2432, 100, 0x1, Int32(10000), typemax(Int64), typemax(Int16), 8, 2^32]
+    vals = [2432, 100, 0x1, Int32(10000), typemax(Int64), typemax(Int16), 8, Int64(2)^32]
     for a_base in vals
         # Only test negative numbers on `a`, since fldmod_by_const requires b > 0.
         @testset for (a, b, f) in Iterators.product((a_base, -a_base), vals, (unsigned, signed))
@@ -14,7 +14,7 @@ using FixedPointDecimals
 end
 
 @testset "div_by_const" begin
-    vals = [2432, 100, 0x1, Int32(10000), typemax(Int64), typemax(Int16), 8, 2^32]
+    vals = [2432, 100, 0x1, Int32(10000), typemax(Int64), typemax(Int16), 8, Int64(2)^32]
     for a_base in vals
         # Only test negative numbers on `a`, since div_by_const requires b > 0.
         @testset for (a, b, f) in Iterators.product((a_base, -a_base), vals, (unsigned, signed))
